requirements.txt
Flask
flask_sqlalchemy
SQLAlchemy
psycopg2
python-dotenv